requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.16
idna==2.9
Kivy==2.0.0
Kivy-Garden==0.1.4
Pillow==7.1.1
Pygments==2.6.1
requests==2.23.0
urllib3==1.25.8
